// apps/sponsor-web/src/types.ts
export interface RegisterData {
  docUrl:               string
  companyName:          string
  businessRegistration: string
  industry:             string
  companySize:          string
  website:              string
  salutation:           string
  contactPerson:        string
  position:             string
  email:                string
  phone:                string
  officeState:          string
  linkedin:             string
  partnershipType:      string
  budget:               string
  objectives:           string
  sustainabilityGoals:  string
  additionalInfo:       string
}